# openapi-formatted.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/doc-types": {
            "post": {
                "operationId": "DocTypeController_create",
                "summary": "\u521b\u5efa\u6587\u4ef6\u7c7b\u578b",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateDocTypeDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "doc-type"
                ]
            }
        },
        "/api/v1/doc-types/list": {
            "get": {
                "operationId": "DocTypeController_findAll",
                "summary": "\u83b7\u53d6\u6587\u4ef6\u7c7b\u578b\u5217\u8868",
                "parameters": [
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "description": "\u9875\u7801",
                        "schema": {
                            "default": 1,
                            "example": 1,
                            "type": "number"
                        }
                    },
                    {
                        "name": "pageSize",
                        "required": false,
                        "in": "query",
                        "description": "\u6bcf\u9875\u6570\u91cf",
                        "schema": {
                            "default": 10,
                            "example": 10,
                            "type": "number"
                        }
                    },
                    {
                        "name": "keyword",
                        "required": false,
                        "in": "query",
                        "description": "\u5173\u952e\u8bcd\u641c\u7d22",
                        "schema": {
                            "example": "\u5408\u540c",
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "required": false,
                        "in": "query",
                        "description": "\u72b6\u6001\uff1a1-\u542f\u7528\uff0c0-\u505c\u7528",
                        "schema": {
                            "example": 1,
                            "enum": [
                                0,
                                1
                            ],
                            "type": "number"
                        }
                    },
                    {
                        "name": "code",
                        "required": false,
                        "in": "query",
                        "description": "\u6587\u4ef6\u7c7b\u578b\u7f16\u7801",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "projectPhase",
                        "required": false,
                        "in": "query",
                        "description": "\u6240\u5c5e\u9879\u76ee\u9636\u6bb5",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "majorCategory",
                        "required": false,
                        "in": "query",
                        "description": "\u6240\u5c5e\u5927\u7c7b",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "minorCategory",
                        "required": false,
                        "in": "query",
                        "description": "\u6240\u5c5e\u5c0f\u7c7b",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "projectType",
                        "required": false,
                        "in": "query",
                        "description": "\u9002\u7528\u9879\u76ee\u7c7b\u578b",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "doc-type"
                ]
            }
        },
...
        "/api/v1/doc-types/export": {
            "get": {
                "operationId": "DocTypeController_exportExcel",
                "summary": "\u5bfc\u51fa\u6570\u636e\u5230Excel",
                "parameters": [
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "description": "\u9875\u7801",
                        "schema": {
                            "default": 1,
                            "example": 1,
                            "type": "number"
                        }
                    },
                    {
                        "name": "pageSize",
                        "required": false,
                        "in": "query",
                        "description": "\u6bcf\u9875\u6570\u91cf",
                        "schema": {
                            "default": 10,
                            "example": 10,
                            "type": "number"
                        }
                    },
                    {
                        "name": "keyword",
                        "required": false,
                        "in": "query",
                        "description": "\u5173\u952e\u8bcd\u641c\u7d22",
                        "schema": {
                            "example": "\u5408\u540c",
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "required": false,
                        "in": "query",
                        "description": "\u72b6\u6001\uff1a1-\u542f\u7528\uff0c0-\u505c\u7528",
                        "schema": {
                            "example": 1,
                            "enum": [
                                0,
                                1
                            ],
                            "type": "number"
                        }
                    },
                    {
                        "name": "code",
                        "required": false,
                        "in": "query",
                        "description": "\u6587\u4ef6\u7c7b\u578b\u7f16\u7801",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "projectPhase",
                        "required": false,
                        "in": "query",
                        "description": "\u6240\u5c5e\u9879\u76ee\u9636\u6bb5",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "majorCategory",
                        "required": false,
                        "in": "query",
                        "description": "\u6240\u5c5e\u5927\u7c7b",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "minorCategory",
                        "required": false,
                        "in": "query",
                        "description": "\u6240\u5c5e\u5c0f\u7c7b",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "projectType",
                        "required": false,
                        "in": "query",
                        "description": "\u9002\u7528\u9879\u76ee\u7c7b\u578b",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "doc-type"
                ]
            }
        },
        "/api/v1/doc-types/import": {
            "post": {
                "operationId": "DocTypeController_importExcel",
                "summary": "Excel\u6279\u91cf\u5bfc\u5165",
                "description": "\u652f\u6301 upsert\uff08\u9ed8\u8ba4\uff09\u3001insertOnly\u3001updateOnly \u4e09\u79cd\u6a21\u5f0f\uff0cdryRun=true \u65f6\u53ea\u6821\u9a8c\u4e0d\u843d\u5e93",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "mode": {
                                        "type": "string",
                                        "enum": [
                                            "upsert",
                                            "insertOnly",
                                            "updateOnly"
                                        ],
                                        "default": "upsert"
                                    },
                                    "dryRun": {
                                        "type": "boolean",
                                        "default": false
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "doc-type"
                ]
            }
        },
        "/api/v1/doc-types/full/{idOrCode}": {
            "get": {
                "operationId": "DocTypeController_getFullInfo",
                "summary": "\u83b7\u53d6\u6587\u4ef6\u7c7b\u578b\u5b8c\u6574\u4fe1\u606f",
                "description": "\u6839\u636e\u6587\u4ef6\u7c7b\u578bID\u6216\u7f16\u7801\uff0c\u8fd4\u56de\u6587\u4ef6\u7c7b\u578b\u4fe1\u606f\u3001\u5173\u952e\u4fe1\u606f\u5b57\u6bb5\u5217\u8868\u3001\u6587\u4ef6\u6a21\u677f/\u793a\u4f8b\u5217\u8868",
                "parameters": [
                    {
                        "name": "idOrCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "doc-type"
                ]
            }
        },
        "/api/v1/doc-types/{id}": {
            "get": {
                "operationId": "DocTypeController_findOne",
                "summary": "\u83b7\u53d6\u6587\u4ef6\u7c7b\u578b\u8be6\u60c5",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "doc-type"
                ]
            },
            "put": {
                "operationId": "DocTypeController_update",
                "summary": "\u66f4\u65b0\u6587\u4ef6\u7c7b\u578b",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateDocTypeDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "doc-type"
                ]
            },
            "delete": {
                "operationId": "DocTypeController_remove",
                "summary": "\u5220\u9664/\u505c\u7528\u6587\u4ef6\u7c7b\u578b",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "doc-type"
                ]
            }
        },
        "/api/v1/doc-types/batch/enable": {
            "post": {
                "operationId": "DocTypeController_batchEnable",
                "summary": "\u6279\u91cf\u542f\u7528",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BatchIdsDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "doc-type"
                ]
            }
        },
        "/api/v1/doc-types/batch/disable": {
            "post": {
                "operationId": "DocTypeController_batchDisable",
                "summary": "\u6279\u91cf\u505c\u7528",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BatchIdsDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "doc-type"
                ]
            }
        },
        "/api/v1/doc-types/batch/delete": {
            "post": {
                "operationId": "DocTypeController_batchDelete",
                "summary": "\u6279\u91cf\u5220\u9664",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BatchIdsDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "doc-type"
                ]
            }
        },
...
    },
    "info": {
        "title": "\u5de5\u7a0b\u54a8\u8be2\u5168\u4e1a\u52a1\u6570\u636e\u5e93\u5e73\u53f0 API",
        "description": "\n## \u6982\u8ff0\n\u5de5\u7a0b\u54a8\u8be2\u5168\u4e1a\u52a1\u6570\u636e\u5e93\u5e73\u53f0\u540e\u7aef\u63a5\u53e3\u6587\u6863 v1.0\n\n## \u8ba4\u8bc1\n\u6240\u6709\u63a5\u53e3\uff08\u9664\u767b\u5f55\u3001\u6ce8\u518c\u3001\u5065\u5eb7\u68c0\u67e5\u5916\uff09\u5747\u9700\u8981 JWT \u8ba4\u8bc1\u3002\n\u8bf7\u5728\u8bf7\u6c42\u5934\u4e2d\u6dfb\u52a0\uff1a`Authorization: Bearer <token>`\n\n## \u54cd\u5e94\u683c\u5f0f\n- \u6210\u529f\uff1a`{ data: T, meta?: { page, pageSize, total, totalPages } }`\n- \u5931\u8d25\uff1a`{ code: string, message: string, traceId: string, details?: any }`\n\n## \u5206\u9875\n- \u9ed8\u8ba4 page=1, pageSize=10\n- \u6700\u5927 pageSize=1000\n    ",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "JWT": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http",
                "description": "\u8f93\u5165 JWT Token \u8fdb\u884c\u8ba4\u8bc1"
            }
        },
        "schemas": {
...
            "CreateDocTypeDto": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "description": "\u6587\u4ef6\u7c7b\u578b\u7f16\u7801\uff08\u7559\u7a7a\u81ea\u52a8\u751f\u6210\uff09",
                        "example": "DT-202601-000001"
                    },
                    "name": {
                        "type": "string",
                        "description": "\u6587\u4ef6\u7c7b\u578b\u540d\u79f0",
                        "example": "\u65bd\u5de5\u5408\u540c"
                    },
                    "projectPhase": {
                        "type": "string",
                        "description": "\u6240\u5c5e\u9879\u76ee\u9636\u6bb5",
                        "example": "\u62db\u6295\u6807\u9636\u6bb5"
                    },
                    "majorCategory": {
                        "type": "string",
                        "description": "\u6240\u5c5e\u5927\u7c7b",
                        "example": "\u5408\u540c\u7c7b"
                    },
                    "minorCategory": {
                        "type": "string",
                        "description": "\u6240\u5c5e\u5c0f\u7c7b",
                        "example": "\u65bd\u5de5\u5408\u540c"
                    },
                    "fileFeature": {
                        "type": "string",
                        "description": "\u6587\u4ef6\u7279\u5f81\u4fe1\u606f\uff08\u7528\u4e8eLLM\u8bc6\u522b\uff09",
                        "example": "\u5305\u542b\u7532\u65b9\u4e59\u65b9\u4fe1\u606f\u3001\u5408\u540c\u91d1\u989d\u3001\u5de5\u671f\u7b49\u5173\u952e\u6761\u6b3e"
                    },
                    "projectType": {
                        "type": "string",
                        "description": "\u9002\u7528\u9879\u76ee\u7c7b\u578b\uff08\u9017\u53f7\u5206\u9694\uff09",
                        "example": "\u623f\u5efa\u5de5\u7a0b,\u5e02\u653f\u5de5\u7a0b"
                    },
                    "region": {
                        "type": "string",
                        "description": "\u9002\u7528\u5730\u533a",
                        "example": "\u5168\u56fd"
                    },
                    "ownerOrg": {
                        "type": "string",
                        "description": "\u9002\u7528\u4e1a\u4e3b",
                        "example": "\u653f\u5e9c\u6295\u8d44\u9879\u76ee"
                    },
                    "bizDescription": {
                        "type": "string",
                        "description": "\u4e1a\u52a1\u8bf4\u660e/\u4f7f\u7528\u573a\u666f",
                        "example": "\u7528\u4e8e\u5ba1\u8ba1\u65bd\u5de5\u5408\u540c\u7684\u5408\u89c4\u6027\u548c\u91d1\u989d\u51c6\u786e\u6027"
                    },
                    "remark": {
                        "type": "string",
                        "description": "\u5907\u6ce8",
                        "example": "\u9700\u8981\u5173\u6ce8\u5408\u540c\u53d8\u66f4\u60c5\u51b5"
                    }
                },
                "required": [
                    "name"
                ]
            },
            "UpdateDocTypeDto": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "description": "\u6587\u4ef6\u7c7b\u578b\u7f16\u7801\uff08\u7559\u7a7a\u81ea\u52a8\u751f\u6210\uff09",
                        "example": "DT-202601-000001"
                    },
                    "name": {
                        "type": "string",
                        "description": "\u6587\u4ef6\u7c7b\u578b\u540d\u79f0",
                        "example": "\u65bd\u5de5\u5408\u540c"
                    },
                    "projectPhase": {
                        "type": "string",
                        "description": "\u6240\u5c5e\u9879\u76ee\u9636\u6bb5",
                        "example": "\u62db\u6295\u6807\u9636\u6bb5"
                    },
                    "majorCategory": {
                        "type": "string",
                        "description": "\u6240\u5c5e\u5927\u7c7b",
                        "example": "\u5408\u540c\u7c7b"
                    },
                    "minorCategory": {
                        "type": "string",
                        "description": "\u6240\u5c5e\u5c0f\u7c7b",
                        "example": "\u65bd\u5de5\u5408\u540c"
                    },
                    "fileFeature": {
                        "type": "string",
                        "description": "\u6587\u4ef6\u7279\u5f81\u4fe1\u606f\uff08\u7528\u4e8eLLM\u8bc6\u522b\uff09",
                        "example": "\u5305\u542b\u7532\u65b9\u4e59\u65b9\u4fe1\u606f\u3001\u5408\u540c\u91d1\u989d\u3001\u5de5\u671f\u7b49\u5173\u952e\u6761\u6b3e"
                    },
                    "projectType": {
                        "type": "string",
                        "description": "\u9002\u7528\u9879\u76ee\u7c7b\u578b\uff08\u9017\u53f7\u5206\u9694\uff09",
                        "example": "\u623f\u5efa\u5de5\u7a0b,\u5e02\u653f\u5de5\u7a0b"
                    },
                    "region": {
                        "type": "string",
                        "description": "\u9002\u7528\u5730\u533a",
                        "example": "\u5168\u56fd"
                    },
                    "ownerOrg": {
                        "type": "string",
                        "description": "\u9002\u7528\u4e1a\u4e3b",
                        "example": "\u653f\u5e9c\u6295\u8d44\u9879\u76ee"
                    },
                    "bizDescription": {
                        "type": "string",
                        "description": "\u4e1a\u52a1\u8bf4\u660e/\u4f7f\u7528\u573a\u666f",
                        "example": "\u7528\u4e8e\u5ba1\u8ba1\u65bd\u5de5\u5408\u540c\u7684\u5408\u89c4\u6027\u548c\u91d1\u989d\u51c6\u786e\u6027"
                    },
                    "remark": {
                        "type": "string",
                        "description": "\u5907\u6ce8",
                        "example": "\u9700\u8981\u5173\u6ce8\u5408\u540c\u53d8\u66f4\u60c5\u51b5"
                    },
                    "status": {
                        "type": "number",
                        "description": "\u72b6\u6001\uff1a1-\u542f\u7528\uff0c0-\u505c\u7528"
                    },
                    "rowVersion": {
                        "type": "number",
                        "description": "\u6570\u636e\u7248\u672c\u53f7\uff08\u4e50\u89c2\u9501\uff09"
                    }
                }
            },
            "BatchIdsDto": {
                "type": "object",
                "properties": {
                    "ids": {
                        "description": "ID\u6570\u7ec4",
                        "type": "array",
                        "items": {
                            "type": "number"
                        }
                    }
                },
                "required": [
                    "ids"
                ]
            },
...
        }
    }
}